+++ src/components/login-top-block/login-top-block.module.scss
@@ -40,4 +40,3 @@
         background-color: #53c41a1f;
     }
 }
-
