+++ src/components/login-top-block/login-top-block.tsx
@@ -1,25 +1,10 @@
 import { Button } from 'antd'
 // eslint-disable-next-line import/named
 import { RouteComponentProps, withRouter } from 'react-router-dom'
-import { useEffect } from 'react'
-import { useDispatch, useSelector } from 'react-redux'
-
-import { getUser, RootState } from '../../services/store/user-slice'
-import { AppDispatch } from '../../services/store/store'
 
 import classes from './login-top-block.module.scss'
 
 const LoginTopBlock = (props: RouteComponentProps) => {
-  const userState = useSelector((state: RootState) => state.users)
-  const dispatch = useDispatch<AppDispatch>()
-  useEffect(() => {
-    console.log('header useeffect')
-    if (!userState.isLoggedIn && window.localStorage.getItem('auth_token')) {
-      console.log('gettting user')
-      dispatch(getUser())
-    }
-  }, [])
-
   return (
     <div className={classes['login-top-block']}>
       <Button
