+++ src/components/user-top-block/index.tsx
@@ -1,0 +1,3 @@
+import UserTopBlock from './user-top-block'
+
+export default UserTopBlock
